% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_border}
\alias{add_border}
\title{Add a Border to the Edge of the Hexagon}
\usage{
add_border(border_width = 0.05, border_col = "black")
}
\arguments{
\item{border_width}{Numeric. Thickness of the border, expressed as the
inverse ratio of the interior of the hex to the full extent of the hex
(must be between \code{0} and \code{1}).}

\item{border_col}{Character. Named R colour or hexadecimal code for the
border around the hex.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Add a border of certain thickness and colour to the hexagon. Should
be called after \link{open_device} and \link{add_hex}, in that order. If desired, make
this the last function before \link{close_device} so that the border overlays all
other elements.
}
\details{
\subsection{Colours}{

Named colour values must be listed in \code{\link[grDevices:colors]{grDevices::colours()}}. Hexadecimal
colour values must be provided with length 6 or 8 and must begin with an
octothorpe (\verb{#}).
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
add_border()
close_device()
}
