% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_hex}
\alias{add_hex}
\title{Add a Hexagon}
\usage{
add_hex(bg_col = "grey")
}
\arguments{
\item{bg_col}{Character. Named R colour or hexadecimal code for the interior
background.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Add a hexagon 'canvas' to which elements can be added.
}
\details{
\subsection{Colours}{

Named colour values must be listed in \code{\link[grDevices:colors]{grDevices::colours()}}. Hexadecimal
colour values must be provided with length 6 or 8 and must begin with an
octothorpe (\verb{#}).
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
add_border()
close_device()
}
