% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_image}
\alias{add_image}
\title{Add an Image}
\usage{
add_image(
  image_object,
  image_x = 0.5,
  image_y = 0.7,
  image_angle = 0,
  image_width = 0.4
)
}
\arguments{
\item{image_object}{Array. A PNG file read in by the user. \code{NULL} for no
image.}

\item{image_x}{Numeric. Image location x-axis.}

\item{image_y}{Numeric. Image location y-axis.}

\item{image_angle}{Numeric. Text rotation in degrees.}

\item{image_width}{Numeric. Image width.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Overlay an image on the hexagon. Image content outside the hexagon will be
clipped. Should be called after \link{open_device} and \link{add_hex}, in that order.
Call this function separately for each image you want to add.
}
\details{
\subsection{Coordinates}{

Coordinates should be provided within the x- and y-axis ranges, which are
both from 0 to 1, giving the centre as x = 0.5 and y = 0.5.
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
add_border()
close_device()
}
