% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_text}
\alias{add_text}
\title{Add Text}
\usage{
add_text(
  text_string = "example",
  text_x = 0.5,
  text_y = 0.3,
  text_angle = 0,
  text_size = 20,
  text_col = "black",
  text_family = "sans",
  text_face = c("plain", "bold", "italic", "bold.italic")
)
}
\arguments{
\item{text_string}{Character. Text to display. \code{NULL} (or an empty string) if
you don't want to place text.}

\item{text_x}{Numeric. Text location x-axis.}

\item{text_y}{Numeric. Text location y-axis.}

\item{text_angle}{Numeric. Rotation of text string in degrees. Positive
values will rotate anticlockwise by the given angle.}

\item{text_size}{Numeric. Text point-size.}

\item{text_col}{Character. Text colour. A named R colour or hexadecimal code.}

\item{text_family}{Character. Name of a font family available on your system.}

\item{text_face}{Character. Font face for the text.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Overlay text on the hexagon. Text outside the hexagon will be clipped. Should
be called after \link{open_device} and \link{add_hex}, in that order. Call this function
separately for each text item that you want to add.
}
\details{
\subsection{Coordinates}{

Coordinates should be provided within the x- and y-axis ranges, which are
both from 0 to 1, giving the centre as x = 0.5 and y = 0.5.
}

\subsection{Colours}{

Named colour values must be listed in \code{\link[grDevices:colors]{grDevices::colours()}}. Hexadecimal
colour values must be provided with length 6 or 8 and must begin with an
octothorpe (\verb{#}).
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
close_device()
}
