% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{close_device}
\alias{close_device}
\title{Close the Device}
\usage{
close_device()
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Clip to the area of the outer hexagon and close the PNG plot device, which
writes to the \code{file_path} specified in \link{open_device}.
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
add_border()
close_device()
}
