% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gex-package.R
\docType{package}
\name{gex-package}
\alias{gex}
\alias{gex-package}
\title{gex: A Dependency-Light Hex-Logo Builder}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A simple interface to create hexagon-shaped logos that help promote your R package or other projects. Uses the 'grid' system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/gex}
  \item Report bugs at \url{https://github.com/matt-dray/gex/bugs}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
