% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbase-package.R
\docType{package}
\name{hexbase-package}
\alias{hexbase}
\alias{hexbase-package}
\title{hexbase: Simple Dependency-Free Hex Logos}
\description{
A simple, dependency-free way to create hexagon-shaped logos to help promote your R package or other projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/hexbase}
  \item Report bugs at \url{https://github.com/matt-dray/hexbase/bugs}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
