% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{open_device}
\alias{open_device}
\title{Open a Device}
\usage{
open_device(file_path)
}
\arguments{
\item{file_path}{Character. Full file path to a .png where the output file
will be saved. The containing directory must already exist.}
}
\value{
Nothing. A graphics device is opened.
}
\description{
Begin a PNG plot device with dimensions matching
\href{https://sticker.how/#type-hexagon}{the Stickers Standard}: 4.39 cm wide by
5.08 cm high (2 by 1.73 inches).
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
image_path <- system.file("img", "Rlogo.png", package = "png")
image_png <- png::readPNG(image_path)
add_image(image_png)
add_text()
add_border()
close_device()
}
