% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_border}
\alias{add_border}
\title{Add a Border to the Edge of the Hexagon}
\usage{
add_border(width = 0.05, col = "black")
}
\arguments{
\item{width}{Numeric. Thickness of the border, expressed as the
inverse ratio of the interior of the hex to the full extent of the hex
(must be between \code{0} and \code{1}).}

\item{col}{Character. Named R colour or hexadecimal code for the
border around the hex.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Add a border of given thickness and colour to the inner edges of hexagon.
}
\details{
\subsection{Order}{

When building a hex, this function should be called after \link{open_device},
\link{add_hex} and any calls to \link{add_text} and \link{add_image} (in that order), and
before \link{close_device}.
}

\subsection{Colours}{

Named colour values must be listed in \code{\link[grDevices:colors]{grDevices::colours()}}. Hexadecimal
colour values must be provided with length 6 or 8 and must begin with an
octothorpe (\verb{#}).
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
img_path <- system.file("img", "Rlogo.png", package = "png")
img_png <- png::readPNG(img_path)
add_image(img_png)
add_text()
add_border()
close_device()
}
\seealso{
Other hex elements: 
\code{\link{add_hex}()}
}
\concept{hex elements}
