% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_hex}
\alias{add_hex}
\title{Add a Hexagon}
\usage{
add_hex(col = "grey")
}
\arguments{
\item{col}{Character. Named R colour or hexadecimal code for the interior
background.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Add a hexagon 'canvas' to which elements can be added.
}
\details{
\subsection{Order}{

When building a hex, this function should be called after \link{open_device}. You can then use
\link{add_text}, \link{add_image} and \link{add_border} (if desired) and finally
\link{close_device}.
}

\subsection{Colours}{

Named colour values must be listed in \code{\link[grDevices:colors]{grDevices::colours()}}. Hexadecimal
colour values must be provided with length 6 or 8 and must begin with an
octothorpe (\verb{#}).
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
img_path <- system.file("img", "Rlogo.png", package = "png")
img_png <- png::readPNG(img_path)
add_image(img_png)
add_text()
add_border()
close_device()
}
\seealso{
Other hex elements: 
\code{\link{add_border}()}
}
\concept{hex elements}
