% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon.R
\name{add_image}
\alias{add_image}
\title{Add an Image to the Hexagon}
\usage{
add_image(img, x = 0.5, y = 0.7, angle = 0, width = 0.4)
}
\arguments{
\item{img}{Array. A PNG or JPEG file read in by the user, most likely
using packages 'png' or 'jpeg'.}

\item{x}{Numeric. Image location on the hexagon's x-axis.}

\item{y}{Numeric. Image location on the hexagon's y-axis.}

\item{angle}{Numeric. Text rotation in degrees.}

\item{width}{Numeric. Image width.}
}
\value{
\code{NULL}. Adds to an existing graphics device.
}
\description{
Overlay an image on the hexagon. Call this function separately for each image
you want to add.
}
\details{
\subsection{Order}{

When building a hex, this function should be called after \link{open_device} and
\link{add_hex}. You can then use further calls to \link{add_image}, \link{add_text} and
\link{add_border} (if desired) and finally \link{close_device}.
}

\subsection{Coordinates}{

Coordinates should be provided within the x- and y-axis ranges, which are
both from 0 to 1, giving the centre as x = 0.5 and y = 0.5.
}
}
\examples{
temp_path <- tempfile(fileext = ".png")
open_device(temp_path)
add_hex()
img_path <- system.file("img", "Rlogo.png", package = "png")
img_png <- png::readPNG(img_path)
add_image(img_png)
add_text()
add_border()
close_device()
}
