% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbase-package.R
\docType{package}
\name{hexbase-package}
\alias{hexbase}
\alias{hexbase-package}
\title{hexbase: Make a Hex Logo}
\description{
Dependency-free way to create a hexagon-shaped logo to help promote your R package or other projects. Work in progress.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/matt-dray/hexbase/bugs}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
