% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbase-package.R
\docType{package}
\name{hexbase-package}
\alias{hexbase}
\alias{hexbase-package}
\title{hexbase: Create Simple Dependency-Free Hex Logos}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A simple interface to create hexagon-shaped logos that help promote your R package or other projects. Uses base R only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/hexbase}
  \item Report bugs at \url{https://github.com/matt-dray/hexbase/bugs}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
