% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{make_hex}
\alias{make_hex}
\title{Create a Hex}
\usage{
make_hex(
  filename,
  bg_col = "white",
  border_col = "#000000",
  border_width = 0.95,
  text = "hexbase",
  text_x = 0,
  text_y = 0,
  text_size = 2
)
}
\arguments{
\item{filename}{Character. Full file path to a .png where the output PNG will
be saved. The containing directory must exist.}

\item{bg_col}{Character. Named R colour or hex code for the interior
background.}

\item{border_col}{Character. Named R colour or hex code for the border around
the hex.x}

\item{border_width}{Numeric. Thickness of the border, expressed as a ratio
of the 'inner' hexagon to 'outer' hexagon (must always be less than 1).}

\item{text}{Character. Text to add to the hex.}

\item{text_x}{Numeric. The x-axis position where the text will be placed.
Defaults to 0 (centre).}

\item{text_y}{Numeric. The y-axis position where the text will be placed.
Defaults to 0 (centre).}

\item{text_size}{Numeric. Size of the text in pixels.}
}
\value{
A new plot.
}
\description{
Create a Hex
}
\details{
Plots a hexagon to the dimensions of the sticker standard given by
<github.com/terinjokes/StickersStandard>: 4.39 cm wide by 5.08 cm high.
}
\examples{
\dontrun{make_hex()}
}
